Create Safe U-Turns
In the proposed solution a blind U-turn is considered and the indication system is designed for
accident prevention when vehicles are coming from both ends of the road that is when the
indication system kicks in, it consists of two LEDs; red and green which indicate the presence
of a vehicle on the either of the turn and no vehicle on the road respectively. The system works
on two main principles which are the Cross-Circuit indicator which provides enhanced
visibility to the driver and power motion sensors which help in the conversion of the kinetic
energy produced by the motion of the vehicle into electric energy which is then stored in the
battery and is used for indication purposes.